+++ ___tests___/set.test.js
@@ -6,19 +6,19 @@ const { set } = require('../index');
 describe('Test fatalError option on set', () => {
   it('Set with append and fetalError', () => {
     const testObject = {};
-    const result = set(testObject, 'stores[{$append}].storeName', 'Amsterdam', {}, { fatalErrorOnCreate: true });
+    const result = set(testObject, 'stores[{$append}].storeName', 'Amsterdam', { fatalErrorOnCreate: true });
     expect(result).toStrictEqual({ stores: [{ storeName: 'Amsterdam' }] });
   });
   it('Set with append and no fetalError', () => {
     const testObject = {};
-    const result = set(testObject, 'stores[{$append}].storeName', 'Amsterdam', {}, { fatalErrorOnCreate: false });
+    const result = set(testObject, 'stores[{$append}].storeName', 'Amsterdam', { fatalErrorOnCreate: false });
     expect(result).toStrictEqual({ stores: [{ storeName: 'Amsterdam' }] });
   });
   it('Set with end and fetalError', () => {
     const testObject = {};
     let errorMessage = null;
     try {
-      result = set(testObject, 'stores[{$end}].storeName', 'Amsterdam', {}, { fatalErrorOnCreate: true });
+      result = set(testObject, 'stores[{$end}].storeName', 'Amsterdam', { fatalErrorOnCreate: true });
     } catch (e) {
       errorMessage = e.message;
     }
@@ -26,14 +26,14 @@ describe('Test fatalError option on set', () => {
   });
   it('Set with end and no fetalError', () => {
     const testObject = {};
-    const result = set(testObject, 'stores[{$end}].storeName', 'Amsterdam', {}, { fatalErrorOnCreate: false });
+    const result = set(testObject, 'stores[{$end}].storeName', 'Amsterdam', { fatalErrorOnCreate: false });
     expect(result).toStrictEqual({});
   });
   it('Set with end to empty array', () => {
     const testObject = { stores: [] };
     let errorMessage = null;
     try {
-      result = set(testObject, 'stores[{$end}].storeName', 'Amsterdam', {}, { fatalErrorOnCreate: true });
+      result = set(testObject, 'stores[{$end}].storeName', 'Amsterdam', { fatalErrorOnCreate: true });
     } catch (e) {
       errorMessage = e.message;
     }
@@ -41,24 +41,24 @@ describe('Test fatalError option on set', () => {
   });
   it('Test wildcard for object', () => {
     const testObject = { wrapper: { key1: {}, key2: {} } };
-    const result = set(testObject, 'wrapper[*].test', true, {}, { fatalErrorOnCreate: false });
+    const result = set(testObject, 'wrapper[*].test', true, { fatalErrorOnCreate: false });
     expect(result).toStrictEqual({ wrapper: { key1: { test: true }, key2: { test: true } } });
   });
   it('Test wildcard for array', () => {
     const testObject = { wrapper: [{ }, { }] };
-    const result = set(testObject, 'wrapper[*].test', true, {}, { fatalErrorOnCreate: false });
+    const result = set(testObject, 'wrapper[*].test', true, { fatalErrorOnCreate: false });
     expect(result).toStrictEqual({ wrapper: [{ test: true }, { test: true }] });
   });
   it('Use of wildcard for non-existing element', () => {
     const testObject = { wrapper: [{}, {}] };
-    const result = set(testObject, 'wrapper2[*].test', true, {}, { fatalErrorOnCreate: false });
+    const result = set(testObject, 'wrapper2[*].test', true, { fatalErrorOnCreate: false });
     expect(result).toStrictEqual({ wrapper: [{}, {}] });
   });
   it('Use of wildcard for non-existing element returns error', () => {
     const testObject = { wrapper: [{}, {}] };
     let errorMessage = null;
     try {
-      result = set(testObject, 'wrapper2[*].test', true, {}, { fatalErrorOnCreate: true });
+      result = set(testObject, 'wrapper2[*].test', true, { fatalErrorOnCreate: true });
     } catch (e) {
       errorMessage = e.message;
     }
@@ -68,7 +68,7 @@ describe('Test fatalError option on set', () => {
     const testObject = { wrapper: [] };
     let errorMessage = null;
     try {
-      result = set(testObject, 'wrapper[*].test', true, {}, { fatalErrorOnCreate: true });
+      result = set(testObject, 'wrapper[*].test', true, { fatalErrorOnCreate: true });
     } catch (e) {
       errorMessage = e.message;
     }
@@ -77,7 +77,7 @@ describe('Test fatalError option on set', () => {
   it('Use of custom function', () => {
     const testObject = [[], [{ test: true }, { test: true }, { test: false }]];
     const customFunction = (element) => element.test === true;
-    const result = set(testObject, '[*][{$Function(customFunction)}].test2', true, { customFunction }, { fatalErrorOnCreate: false });
+    const result = set(testObject, '[*][{$Function(customFunction)}].test2', true, { functions: { customFunction }, fatalErrorOnCreate: false });
     expect(result).toStrictEqual(
       [[], [{ test: true, test2: true }, { test: true, test2: true }, { test: false }]],
     );