+++ src/clients/json/index.js
@@ -11,7 +11,7 @@ const getPath = require('../../handlers/getPath');
 const getPaths = require('../../handlers/getPaths');
 const getAllPaths = require('../../handlers/getAllPaths');
 const chop = require('../../handlers/chop');
-const mergeObjects = require('../../handlers/mergeObjects');
+const mergeSettings = require('./src/mergeSettings');
 
 const SettingsLoader = require('../../services/settingsLoader');
 
@@ -22,7 +22,7 @@ class Json {
    * @param {Object} settings - object with settings
    * @param {Object} functions - object of functions that can be called within query.
    */
-  constructor(object, settings, functions) {
+  constructor(object, settings) {
     const settingsLoader = new SettingsLoader({ settings });
     this.settings = settingsLoader.settings;
     if (this.settings.unlinkInputObject) {
@@ -30,128 +30,118 @@ class Json {
     } else {
       this.object = settingsLoader.makeJson(object);
     }
-    this.functions = settingsLoader.makeObject(functions);
   }
 
   /**
    * Retrieves single value from objects specified query path
    * @param {any} path - string or array representation of path to set.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {any} returns value found at specified path, in case that multiple logical checks
    * satisfy the first element will be returned
    */
-  getOne(path, functions, settings) {
+  getOne(path, settings) {
     return getOne(
-      this.object, path, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path,
+      mergeSettings(this.settings, settings),
     );
   }
 
   /**
    * Finds single value and path from objects specified query path
    * @param {any} path - string or array representation of path to set.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {any} returns value and resolved path found at specified path,
    * in case that multiple logical checks satisfy the first element will be returned
    */
-  findOne(path, functions, settings) {
+  findOne(path, settings) {
     return findOne(
-      this.object, path, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path,
+      mergeSettings(this.settings, settings),
     );
   }
 
   /**
    * Retrieves resolved path from objects specified path
    * @param {any} path - string or array representation of path to set.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {any} returns resolved path found at specified path,
    * in case that multiple logical checks satisfy the first element will be returned
    */
-  getPath(path, functions, settings) {
+  getPath(path, settings) {
     return getPath(
-      this.object, path, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path,
+      mergeSettings(this.settings, settings),
     );
   }
 
   /**
    * Retrieves values from objects specified query path
    * @param {any} path - string or array representation of path to set.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {Array} returns array of values that match the specified path with logical checks
    */
-  get(path, functions, settings) {
+  get(path, settings) {
     return get(
-      this.object, path, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path,
+      mergeSettings(this.settings, settings),
     );
   }
 
   /**
    * Retrieves all values from objects specified query path
    * @param {any} path - string or array representation of path to set.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {Array} returns array of values that match the specified path with logical checks
    */
-  getAll(path, functions, settings) {
+  getAll(path, settings) {
     return getAll(
-      this.object, path, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path,
+      mergeSettings(this.settings, settings),
     );
   }
 
   /**
    * Finds values and paths from objects specified query path
    * @param {any} path - string or array representation of path to set.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {Array} returns array of objects with value/path property
    * that match the specified path with logical checks
    */
-  find(path, functions, settings) {
+  find(path, settings) {
     return find(
-      this.object, path, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path,
+      mergeSettings(this.settings, settings),
     );
   }
 
   /**
    * Finds all values and paths from objects specified query path
    * @param {any} path - string or array representation of path to set.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {Array} returns array of objects with value/path property
    * that match the specified path with logical checks
    */
-  findAll(path, functions, settings) {
+  findAll(path, settings) {
     return findAll(
-      this.object, path, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path,
+      mergeSettings(this.settings, settings),
     );
   }
 
   /**
    * Retrieves resolved paths from objects specified path
    * @param {any} path - string or array representation of path to set.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {Array} returns array of paths that match the specified path with logical checks
    */
-  getPaths(path, functions, settings) {
+  getPaths(path, settings) {
     return getPaths(
-      this.object, path, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path,
+      mergeSettings(this.settings, settings),
     );
   }
 
   /**
    * Retrieves all resolved paths from objects specified path
    * @param {any} path - string or array representation of path to set.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {Array} returns array of paths that match the specified path with logical checks
    */
-  getAllPaths(path, functions, settings) {
+  getAllPaths(path, settings) {
     return getAllPaths(
-      this.object, path, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path,
+      mergeSettings(this.settings, settings),
     );
   }
 
@@ -159,14 +149,13 @@ class Json {
    * Sets single value on specified path
    * @param {any} path - string or array representation of path to set.
    * @param {any} val - value to be set at specified path.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {object} object with newly set path in case that multiple logical checks
    * satisfy the first element will be set.
    */
-  setOne(path, val, functions, settings) {
+  setOne(path, val, settings) {
     return setOne(
-      this.object, path, val, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path, val,
+      mergeSettings(this.settings, settings),
     );
   }
 
@@ -174,14 +163,13 @@ class Json {
    * Sets values on specified path
    * @param {any} path - string or array representation of path to set.
    * @param {any} val - value to be set at specified path.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {object} object with newly set path in case that multiple logical checks
    * satisfy the first element will be set.
    */
-  set(path, val, functions, settings) {
+  set(path, val, settings) {
     return set(
-      this.object, path, val, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path, val,
+      mergeSettings(this.settings, settings),
     );
   }
 
@@ -189,14 +177,13 @@ class Json {
    * Sets all values on specified path
    * @param {any} path - string or array representation of path to set.
    * @param {any} val - value to be set at specified path.
-   * @param {Object} functions - object of functions that can be called within query.
    * @returns {object} object with newly set path in case that multiple logical checks
    * satisfy the first element will be set.
    */
-  setAll(path, val, functions, settings) {
+  setAll(path, val, settings) {
     return setAll(
-      this.object, path, val, mergeObjects([this.functions, functions]),
-      mergeObjects([this.settings, settings]),
+      this.object, path, val,
+      mergeSettings(this.settings, settings),
     );
   }