+++ src/handlers/set/index.js
@@ -5,13 +5,12 @@ const Builder = require('../../services/builder');
  * @param {object} obj - object
  * @param {any} path - string or array representation of path to set.
  * @param {any} val - value to be set at specified path.
- * @param {Object} functions - object of functions that can be called within query.
  * @param {Object} settings - object with settings.
  * @returns {object} object with newly set path in case that multiple logical checks
  * satisfy the first element will be set.
  */
-const set = (object, path, value, functions, settings) => {
-  const builder = new Builder({ functions, settings });
+const set = (object, path, value, settings) => {
+  const builder = new Builder({ settings });
   return builder.build(object, path, value);
 };